support resize embs

+++ infer.py
@@ -15,7 +15,7 @@ if __name__ == '__main__':
     tokenizer: MossTokenizer
     tokenizer, config, _,_ = dataHelper.load_tokenizer_and_config(tokenizer_class_name=MossTokenizer, config_class_name=MossConfig,config_kwargs={"torch_dtype": "float16"})
 
-    pl_model = MyTransformer(config=config, model_args=model_args)
+    pl_model = MyTransformer(config=config, model_args=model_args, torch_dtype=torch.float16,)
     model = pl_model.get_llm_model()
     model.eval().half().cuda()
